# FooBarServiceTracker/FooBarServiceTracker.Api/Infrastructure/Dtos/ServiceDto.cs
namespace FooBarServiceTracker.Api.Infrastructure.Dtos
{
    public class ServiceDto
    {
        public string Name { get; init; } = string.Empty;
        public int? Port { get; init; }
        public string? Maintainer { get; init; }
        public Dictionary<string, string>? Labels { get; init; } 
    }
}